Added in Interest module to practice Object Comparison
